+++ ExtLibs/Controls/WindDir.cs
@@ -19,8 +19,8 @@ namespace MissionPlanner.Controls
 
         }
 
-        const float rad2deg = (float)(180 / Math.PI);
-        const float deg2rad = (float)(1.0 / rad2deg);
+        const double rad2deg = (180 / Math.PI);
+        const double deg2rad = (1.0 / rad2deg);
         double _direction = 0;
         double _speed = 0;
