+++ ExtLibs/Grid/Grid.cs
@@ -15,8 +15,8 @@ namespace MissionPlanner
     {
         public static MissionPlanner.Plugin.PluginHost Host2;
 
-        const float rad2deg = (float)(180 / Math.PI);
-        const float deg2rad = (float)(1.0 / rad2deg);
+        const double rad2deg = (180 / Math.PI);
+        const double deg2rad = (1.0 / rad2deg);
 
         public struct linelatlng
         {
