+++ GCSViews/ConfigurationView/ConfigHWCompass.cs
@@ -11,8 +11,8 @@ namespace MissionPlanner.GCSViews.ConfigurationView
 {
     public partial class ConfigHWCompass : UserControl, IActivate
     {
-        private const double rad2deg = (float) (180/Math.PI);
-        private const double deg2rad = (float) (1.0/rad2deg);
+        private const double rad2deg = (180/Math.PI);
+        private const double deg2rad = (1.0/rad2deg);
         private const int THRESHOLD_OFS_RED = 600;
         private const int THRESHOLD_OFS_YELLOW = 400;
         private bool startup;
