+++ GCSViews/FlightData.cs
@@ -90,9 +90,9 @@ namespace MissionPlanner.GCSViews
 
         List<PointLatLng> trackPoints = new List<PointLatLng>();
 
-        const double rad2deg = (float) (180/Math.PI);
+        const double rad2deg = (180/Math.PI);
 
-        const double deg2rad = (float) (1.0/rad2deg);
+        const double deg2rad = (1.0/rad2deg);
 
         public static HUD myhud;
         public static myGMAP mymap;
